fix np.int not exist problem

+++ matterhorn/data/aedat.py
@@ -91,7 +91,7 @@ class AEDAT(EventDataset2d):
         @return:
             data_tpyx: np.ndarray 分为t,p,y,x的数据，形状为[n, 4]
         """
-        res = np.zeros((data.shape[0] // 2, 4), dtype = np.int)
+        res = np.zeros((data.shape[0] // 2, 4), dtype = "int32")
         xyp = data[::2]
         t = data[1::2]
         res[:, 0] = t
@@ -257,7 +257,7 @@ class CIFAR10DVS(AEDAT):
         """
         list_filename = os.path.join(self.processed_folder, "__main__.csv")
         if os.path.isfile(list_filename):
-            file_list = np.loadtxt(list_filename, dtype = np.int, delimiter = ",")
+            file_list = np.loadtxt(list_filename, dtype = "int32", delimiter = ",")
             return file_list
         self.unzip()
         os.makedirs(self.processed_folder, exist_ok = True)
@@ -273,7 +273,7 @@ class CIFAR10DVS(AEDAT):
                 self.save_event_data(file_idx, event_data)
                 file_list.append([file_idx, label, 1 if self.is_train(label, file_idx % aedat_file_count) else 0])
                 file_idx += 1
-        file_list = np.array(file_list, dtype = np.int)
+        file_list = np.array(file_list, dtype = "int32")
         np.savetxt(list_filename, file_list, fmt = "%d", delimiter = ",")
         return file_list
 
@@ -388,7 +388,7 @@ class DVS128Gesture(AEDAT):
         """
         list_filename = os.path.join(self.processed_folder, "__main__.csv")
         if os.path.isfile(list_filename):
-            file_list = np.loadtxt(list_filename, dtype = np.int, delimiter = ",")
+            file_list = np.loadtxt(list_filename, dtype = "int32", delimiter = ",")
             return file_list
         self.unzip()
         aedat_file_dir = os.path.join(self.extracted_folder, "DvsGesture")
@@ -409,7 +409,7 @@ class DVS128Gesture(AEDAT):
             event_data = self.data_2_tpyx(raw_data)
             event_data[:, 2] = self.original_size[2] - 1 - event_data[:, 2]
             event_data[:, 3] = self.original_size[3] - 1 - event_data[:, 3]
-            class_info = np.loadtxt(os.path.join(aedat_file_dir, filename.replace(".aedat", "_labels.csv")), dtype = np.int, delimiter = ",", skiprows = 1)
+            class_info = np.loadtxt(os.path.join(aedat_file_dir, filename.replace(".aedat", "_labels.csv")), dtype = "int32", delimiter = ",", skiprows = 1)
             for label, start, end in class_info:
                 data = event_data[(event_data[:, 0] >= start) & (event_data[:, 0] < end)]
                 data[:, 0] = data[:, 0] - start
@@ -419,6 +419,6 @@ class DVS128Gesture(AEDAT):
                 self.save_event_data(file_idx, event_data)
                 file_list.append([file_idx, label - 1, 1 if is_train else 0])
                 file_idx += 1
-        file_list = np.array(file_list, dtype = np.int)
+        file_list = np.array(file_list, dtype = "int32")
         np.savetxt(list_filename, file_list, fmt = "%d", delimiter = ",")
         return file_list
